Refactor Batch datasources into relevant offline store files
