+++ sdk/python/feast/data_source.py
@@ -16,10 +16,9 @@
 import enum
 from typing import Callable, Dict, Iterable, Optional, Tuple
 
-from pyarrow.parquet import ParquetFile
 from tenacity import retry, retry_unless_exception_type, wait_exponential
 
-from feast import type_map
+from feast import BigQuerySource, FileSource, type_map
 from feast.data_format import FileFormat, StreamFormat
 from feast.errors import (
     DataSourceNotFoundException,
@@ -41,151 +40,6 @@ class SourceType(enum.Enum):
     BATCH_BIGQUERY = 2
     STREAM_KAFKA = 3
     STREAM_KINESIS = 4
-
-
-class FileOptions:
-    """
-    DataSource File options used to source features from a file
-    """
-
-    def __init__(
-        self, file_format: Optional[FileFormat], file_url: Optional[str],
-    ):
-        self._file_format = file_format
-        self._file_url = file_url
-
-    @property
-    def file_format(self):
-        """
-        Returns the file format of this file
-        """
-        return self._file_format
-
-    @file_format.setter
-    def file_format(self, file_format):
-        """
-        Sets the file format of this file
-        """
-        self._file_format = file_format
-
-    @property
-    def file_url(self):
-        """
-        Returns the file url of this file
-        """
-        return self._file_url
-
-    @file_url.setter
-    def file_url(self, file_url):
-        """
-        Sets the file url of this file
-        """
-        self._file_url = file_url
-
-    @classmethod
-    def from_proto(cls, file_options_proto: DataSourceProto.FileOptions):
-        """
-        Creates a FileOptions from a protobuf representation of a file option
-
-        args:
-            file_options_proto: a protobuf representation of a datasource
-
-        Returns:
-            Returns a FileOptions object based on the file_options protobuf
-        """
-        file_options = cls(
-            file_format=FileFormat.from_proto(file_options_proto.file_format),
-            file_url=file_options_proto.file_url,
-        )
-        return file_options
-
-    def to_proto(self) -> DataSourceProto.FileOptions:
-        """
-        Converts an FileOptionsProto object to its protobuf representation.
-
-        Returns:
-            FileOptionsProto protobuf
-        """
-
-        file_options_proto = DataSourceProto.FileOptions(
-            file_format=(
-                None if self.file_format is None else self.file_format.to_proto()
-            ),
-            file_url=self.file_url,
-        )
-
-        return file_options_proto
-
-
-class BigQueryOptions:
-    """
-    DataSource BigQuery options used to source features from BigQuery query
-    """
-
-    def __init__(self, table_ref: Optional[str], query: Optional[str]):
-        self._table_ref = table_ref
-        self._query = query
-
-    @property
-    def query(self):
-        """
-        Returns the BigQuery SQL query referenced by this source
-        """
-        return self._query
-
-    @query.setter
-    def query(self, query):
-        """
-        Sets the BigQuery SQL query referenced by this source
-        """
-        self._query = query
-
-    @property
-    def table_ref(self):
-        """
-        Returns the table ref of this BQ table
-        """
-        return self._table_ref
-
-    @table_ref.setter
-    def table_ref(self, table_ref):
-        """
-        Sets the table ref of this BQ table
-        """
-        self._table_ref = table_ref
-
-    @classmethod
-    def from_proto(cls, bigquery_options_proto: DataSourceProto.BigQueryOptions):
-        """
-        Creates a BigQueryOptions from a protobuf representation of a BigQuery option
-
-        Args:
-            bigquery_options_proto: A protobuf representation of a DataSource
-
-        Returns:
-            Returns a BigQueryOptions object based on the bigquery_options protobuf
-        """
-
-        bigquery_options = cls(
-            table_ref=bigquery_options_proto.table_ref,
-            query=bigquery_options_proto.query,
-        )
-
-        return bigquery_options
-
-    def to_proto(self) -> DataSourceProto.BigQueryOptions:
-        """
-        Converts an BigQueryOptionsProto object to its protobuf representation.
-
-        Returns:
-            BigQueryOptionsProto protobuf
-        """
-
-        bigquery_options_proto = DataSourceProto.BigQueryOptions(
-            table_ref=self.table_ref, query=self.query,
-        )
-
-        return bigquery_options_proto
 
 
 class KafkaOptions:
@@ -555,229 +409,6 @@ class DataSource:
         Get the list of column names and raw column types
         """
         raise NotImplementedError
-
-
-class FileSource(DataSource):
-    def __init__(
-        self,
-        event_timestamp_column: Optional[str] = "",
-        file_url: Optional[str] = None,
-        path: Optional[str] = None,
-        file_format: FileFormat = None,
-        created_timestamp_column: Optional[str] = "",
-        field_mapping: Optional[Dict[str, str]] = None,
-        date_partition_column: Optional[str] = "",
-    ):
-        """Create a FileSource from a file containing feature data. Only Parquet format supported.
-
-        Args:
-
-            path: File path to file containing feature data. Must contain an event_timestamp column, entity columns and
-                feature columns.
-            event_timestamp_column: Event timestamp column used for point in time joins of feature values.
-            created_timestamp_column (optional): Timestamp column when row was created, used for deduplicating rows.
-            file_url: [Deprecated] Please see path
-            file_format (optional): Explicitly set the file format. Allows Feast to bypass inferring the file format.
-            field_mapping: A dictionary mapping of column names in this data source to feature names in a feature table
-                or view. Only used for feature columns, not entities or timestamp columns.
-
-        Examples:
-            >>> FileSource(path="/data/my_features.parquet", event_timestamp_column="datetime")
-        """
-        if path is None and file_url is None:
-            raise ValueError(
-                'No "path" argument provided. Please set "path" to the location of your file source.'
-            )
-        if file_url:
-            from warnings import warn
-
-            warn(
-                'Argument "file_url" is being deprecated. Please use the "path" argument.'
-            )
-        else:
-            file_url = path
-
-        self._file_options = FileOptions(file_format=file_format, file_url=file_url)
-
-        super().__init__(
-            event_timestamp_column,
-            created_timestamp_column,
-            field_mapping,
-            date_partition_column,
-        )
-
-    def __eq__(self, other):
-        if not isinstance(other, FileSource):
-            raise TypeError("Comparisons should only involve FileSource class objects.")
-
-        return (
-            self.file_options.file_url == other.file_options.file_url
-            and self.file_options.file_format == other.file_options.file_format
-            and self.event_timestamp_column == other.event_timestamp_column
-            and self.created_timestamp_column == other.created_timestamp_column
-            and self.field_mapping == other.field_mapping
-        )
-
-    @property
-    def file_options(self):
-        """
-        Returns the file options of this data source
-        """
-        return self._file_options
-
-    @file_options.setter
-    def file_options(self, file_options):
-        """
-        Sets the file options of this data source
-        """
-        self._file_options = file_options
-
-    @property
-    def path(self):
-        """
-        Returns the file path of this feature data source
-        """
-        return self._file_options.file_url
-
-    def to_proto(self) -> DataSourceProto:
-        data_source_proto = DataSourceProto(
-            type=DataSourceProto.BATCH_FILE,
-            field_mapping=self.field_mapping,
-            file_options=self.file_options.to_proto(),
-        )
-
-        data_source_proto.event_timestamp_column = self.event_timestamp_column
-        data_source_proto.created_timestamp_column = self.created_timestamp_column
-        data_source_proto.date_partition_column = self.date_partition_column
-
-        return data_source_proto
-
-    def validate(self, config: RepoConfig):
-        # TODO: validate a FileSource
-        pass
-
-    @staticmethod
-    def source_datatype_to_feast_value_type() -> Callable[[str], ValueType]:
-        return type_map.pa_to_feast_value_type
-
-    def get_table_column_names_and_types(
-        self, config: RepoConfig
-    ) -> Iterable[Tuple[str, str]]:
-        schema = ParquetFile(self.path).schema_arrow
-        return zip(schema.names, map(str, schema.types))
-
-
-class BigQuerySource(DataSource):
-    def __init__(
-        self,
-        event_timestamp_column: Optional[str] = "",
-        table_ref: Optional[str] = None,
-        created_timestamp_column: Optional[str] = "",
-        field_mapping: Optional[Dict[str, str]] = None,
-        date_partition_column: Optional[str] = "",
-        query: Optional[str] = None,
-    ):
-        self._bigquery_options = BigQueryOptions(table_ref=table_ref, query=query)
-
-        super().__init__(
-            event_timestamp_column,
-            created_timestamp_column,
-            field_mapping,
-            date_partition_column,
-        )
-
-    def __eq__(self, other):
-        if not isinstance(other, BigQuerySource):
-            raise TypeError(
-                "Comparisons should only involve BigQuerySource class objects."
-            )
-
-        return (
-            self.bigquery_options.table_ref == other.bigquery_options.table_ref
-            and self.bigquery_options.query == other.bigquery_options.query
-            and self.event_timestamp_column == other.event_timestamp_column
-            and self.created_timestamp_column == other.created_timestamp_column
-            and self.field_mapping == other.field_mapping
-        )
-
-    @property
-    def table_ref(self):
-        return self._bigquery_options.table_ref
-
-    @property
-    def query(self):
-        return self._bigquery_options.query
-
-    @property
-    def bigquery_options(self):
-        """
-        Returns the bigquery options of this data source
-        """
-        return self._bigquery_options
-
-    @bigquery_options.setter
-    def bigquery_options(self, bigquery_options):
-        """
-        Sets the bigquery options of this data source
-        """
-        self._bigquery_options = bigquery_options
-
-    def to_proto(self) -> DataSourceProto:
-        data_source_proto = DataSourceProto(
-            type=DataSourceProto.BATCH_BIGQUERY,
-            field_mapping=self.field_mapping,
-            bigquery_options=self.bigquery_options.to_proto(),
-        )
-
-        data_source_proto.event_timestamp_column = self.event_timestamp_column
-        data_source_proto.created_timestamp_column = self.created_timestamp_column
-        data_source_proto.date_partition_column = self.date_partition_column
-
-        return data_source_proto
-
-    def validate(self, config: RepoConfig):
-        if not self.query:
-            from google.api_core.exceptions import NotFound
-            from google.cloud import bigquery
-
-            client = bigquery.Client()
-            try:
-                client.get_table(self.table_ref)
-            except NotFound:
-                raise DataSourceNotFoundException(self.table_ref)
-
-    def get_table_query_string(self) -> str:
-        """Returns a string that can directly be used to reference this table in SQL"""
-        if self.table_ref:
-            return f"`{self.table_ref}`"
-        else:
-            return f"({self.query})"
-
-    @staticmethod
-    def source_datatype_to_feast_value_type() -> Callable[[str], ValueType]:
-        return type_map.bq_to_feast_value_type
-
-    def get_table_column_names_and_types(
-        self, config: RepoConfig
-    ) -> Iterable[Tuple[str, str]]:
-        from google.cloud import bigquery
-
-        client = bigquery.Client()
-        if self.table_ref is not None:
-            table_schema = client.get_table(self.table_ref).schema
-            if not isinstance(table_schema[0], bigquery.schema.SchemaField):
-                raise TypeError("Could not parse BigQuery table schema.")
-
-            name_type_pairs = [(field.name, field.field_type) for field in table_schema]
-        else:
-            bq_columns_query = f"SELECT * FROM ({self.query}) LIMIT 1"
-            queryRes = client.query(bq_columns_query).result()
-            name_type_pairs = [
-                (schema_field.name, schema_field.field_type)
-                for schema_field in queryRes.schema
-            ]
-
-        return name_type_pairs
 
 
 class KafkaSource(DataSource):
