+++ sdk/python/feast/feature_table.py
@@ -20,13 +20,8 @@ from google.protobuf.duration_pb2 import Duration
 from google.protobuf.json_format import MessageToDict, MessageToJson
 from google.protobuf.timestamp_pb2 import Timestamp
 
-from feast.data_source import (
-    BigQuerySource,
-    DataSource,
-    FileSource,
-    KafkaSource,
-    KinesisSource,
-)
+from feast import BigQuerySource, FileSource
+from feast.data_source import DataSource, KafkaSource, KinesisSource
 from feast.feature import Feature
 from feast.loaders import yaml as feast_yaml
 from feast.protos.feast.core.FeatureTable_pb2 import FeatureTable as FeatureTableProto
