+++ sdk/python/tests/example_feature_repo_2.py
@@ -1,7 +1,6 @@
 from google.protobuf.duration_pb2 import Duration
 
-from feast import Entity, Feature, FeatureView, ValueType
-from feast.data_source import FileSource
+from feast import Entity, Feature, FeatureView, FileSource, ValueType
 
 driver_hourly_stats = FileSource(
     path="%PARQUET_PATH%",  # placeholder to be replaced by the test
